refactor: StoreInfoView SRP 원칙에 따라 View 분리 완료
