+++ river/Starbucks_Clone/Starbucks_Clone/Sources/Screen/StoreInfo/StoreInfoView.swift
@@ -36,263 +36,6 @@ struct StoreInfoView: View {
     }
 }
 
-fileprivate struct HeaderSegmentView: View {
-    @Namespace private var underlineSegmentedBar
-    @Binding private var tabState: StoreInfoTabState
-    
-    init(tabState: Binding<StoreInfoTabState>) {
-        self._tabState = tabState
-    }
-    
-    fileprivate var body: some View {
-        HStack {
-            ForEach(StoreInfoTabState.allCases) { state in
-                Button {
-                    withAnimation {
-                        tabState = state
-                    }
-                } label: {
-                    VStack(spacing: 4) {
-                        Text(state.title)
-                            .font(.mainTextSemiBold24)
-                            .foregroundStyle(.black)
-                            .frame(minWidth: 220)
-                            .frame(height: 43)
-                        
-                        if tabState == state {
-                            Color.brown
-                                .frame(width: 120, height: 3)
-                                /// 두 개의 뷰 간 애니메이션 자연스럽게 연결하는 수정자
-                                /// - id: 두 뷰가 같은 것처럼 연동하는 식별자
-                                /// - namespace: 그룹화
-                                .matchedGeometryEffect(
-                                    id: "underline",
-                                    in: underlineSegmentedBar
-                                )
-                        } else {
-                            Color.clear.frame(height: 3)
-                        }
-                    }
-                    .frame(maxWidth: .infinity)
-                }
-            }
-        }
-    }
-}
-
-// MARK: 매장 찾기&길찾기를 바인딩하는 뷰
-fileprivate struct StoreDirectionView: View {
-    @Namespace private var underlineSegmentedBar
-    @Bindable private var storeInfoViewModel: StoreInfoViewModel
-    @Binding private var tabState: StoreInfoTabState
-    
-    init(
-        storeInfoViewModel: StoreInfoViewModel,
-        tabState: Binding<StoreInfoTabState>
-    ) {
-        self.storeInfoViewModel = storeInfoViewModel
-        self._tabState = tabState
-    }
-    
-    fileprivate var body: some View {
-        Group {
-            // FIXME: 개발을 위한 조건문 주석
-//            if tabState == .findStore {
-//                MapView(
-//                    viewModel: findStoreViewModel,
-//                    locationManager: locationManager
-//                )
-//            } else if tabState == .directions {
-//                FindLocationView()
-//            }
-            VStack {
-                StoreDirectionContentView(storeInfoViewModel: storeInfoViewModel)
-            }
-        }
-        .matchedGeometryEffect(
-            id: "underline",
-            in: underlineSegmentedBar
-        )
-    }
-}
-
-// MARK: 출발-도착지 입력 뷰
-fileprivate struct StoreDirectionContentView: View {
-    @Bindable private var storeInfoViewModel: StoreInfoViewModel
-    @State private var storeSearchState = StoreSearchState()
-    
-    init(
-        storeInfoViewModel: StoreInfoViewModel
-    ) {
-        self.storeInfoViewModel = storeInfoViewModel
-    }
-    
-    fileprivate var body: some View {
-        VStack {
-            Group {
-                startedLocationInput
-                
-                Spacer().frame(height: 13)
-                
-                finishedLocationInput
-                
-                Spacer().frame(height: 18)
-                
-                FindLocationBtnView()
-                
-                Spacer().frame(height: 28)
-            }
-            .padding(.horizontal, 31)
-            
-            FindStoreListView(
-                storeInfoViewModel: storeInfoViewModel,
-                storeSearchState: $storeSearchState
-            )
-        }
-    }
-    
-    private var startedLocationInput: some View {
-        HStack {
-            Text("출발")
-                .font(.mainTextSemiBold16)
-                .foregroundStyle(Color(.black03))
-            
-            Spacer().frame(width: 15)
-            
-            Button {
-                Task {
-                    await storeInfoViewModel.getCurrentLocationAddress()
-                }
-                if let currentAddress = storeInfoViewModel.currentAddress {
-                    storeSearchState.startAddress = currentAddress
-                }
-            } label: {
-                Text("현재위치")
-                    .font(.mainTextSemiBold16)
-                    .foregroundStyle(Color(.white01))
-                    .padding(.horizontal, 6)
-            }
-            .frame(height: 30)
-            .background(Color(.brown01))
-            .clipShape(RoundedRectangle(cornerRadius: 6))
-            
-            TextField("출발지 입력", text: $storeSearchState.startAddress)
-                .frame(height: 30)
-                .textFieldStyle(.roundedBorder)
-            
-            Spacer().frame(width: 8)
-            
-            Button {
-                if storeSearchState.startAddress != "" {
-                    Task {
-                        await storeInfoViewModel.addressSearchWithKakao(
-                            storeSearchState.startAddress
-                        )
-                    }
-                    storeSearchState.searchType = .startedLocation
-                } else {
-                    print("출발 장소를 입력해 주세요")
-                }
-            } label: {
-                Image(.searchIcon)
-                    .resizable()
-                    .frame(width: 20, height: 20)
-            }
-        }
-    }
-    
-    private var finishedLocationInput: some View {
-        HStack {
-            Text("도착")
-                .font(.mainTextSemiBold16)
-                .foregroundStyle(Color(.black03))
-            
-            Spacer().frame(width: 15)
-            
-            TextField("출발지 입력", text: $storeSearchState.finishAddress)
-                .frame(height: 30)
-                .textFieldStyle(.roundedBorder)
-            
-            Spacer().frame(width: 8)
-            
-            Button {
-                if storeSearchState.finishAddress != "" {
-                    Task {
-                        await storeInfoViewModel.addressSearchWithKakao(
-                            storeSearchState.finishAddress
-                        )
-                    }
-                    storeSearchState.searchType = .finishedLocation
-                } else {
-                    print("출발 장소를 입력해 주세요")
-                }
-            } label: {
-                Image(.searchIcon)
-                    .resizable()
-                    .frame(width: 20, height: 20)
-            }
-
-        }
-    }
-}
-
-fileprivate struct FindLocationBtnView: View {
-    fileprivate var body: some View {
-        Button {
-            print("경로 찾기")
-        } label: {
-            Text("경로 찾기")
-                .font(.mainTextMedium16)
-                .foregroundStyle(Color.white)
-        }
-        .frame(width: 375, height: 38)
-        .background(Color(.green00))
-        .clipShape(RoundedRectangle(cornerRadius: 10))
-    }
-}
-
-fileprivate struct FindStoreListView: View {
-    @Bindable private var storeInfoViewModel: StoreInfoViewModel
-    @Binding private var storeSearchState: StoreSearchState
-    
-    init(
-        storeInfoViewModel: StoreInfoViewModel,
-        storeSearchState: Binding<StoreSearchState>
-    ) {
-        self.storeInfoViewModel = storeInfoViewModel
-        self._storeSearchState = storeSearchState
-    }
-    
-    fileprivate var body: some View {
-        List(storeInfoViewModel.searchPlaceList, id: \.id) { place in
-            Button {
-                switch storeSearchState.searchType {
-                case .startedLocation:
-                    storeSearchState.startAddress = place.name
-                case .finishedLocation:
-                    storeSearchState.finishAddress = place.name
-                }
-            } label: {
-                searchPlaceResultView(place)
-            }
-
-        }
-        .listStyle(.plain)
-    }
-    
-    private func searchPlaceResultView(_ place: StoreInfoModel) -> some View {
-        return VStack(alignment: .leading, spacing: 8) {
-            Text("\(place.name)")
-                .font(.mainTextMedium16)
-                .foregroundStyle(Color(.black01))
-            
-            Text("\(place.address)")
-                .font(.mainTextSemiBold14)
-                .foregroundStyle(Color(.gray04))
-        }
-    }
-}
-
 #Preview {
     StoreInfoView(storeInfoViewModel: .init(locationManager: .shared))
 }
